image_processing_backend/requirements.txt
fastapi
uvicorn
celery
redis
pillow
psycopg2
pandas
python-multipart
sqlalchemy
alembic
